HUGE MAP UPDATE
To repeat an entity animation, no need to create a different name in entities.json but add a suffix in the instantiation (in the map json). To add name parameters using the suffix, simply use an underscore: _. Before the underscore is the entity name. After the underscore are the name parameters. For now, there are only 2 name parameters: the repeat parameter and the repeat direction parameter.
The repeat parameter is used as the following for example: x12
This will cause the program to repeat 12 times the animation of the entity. This parameter always starts with a 'x'.
The repeat direction parameter is used as the following for example: H
This will cause the program to repeat horizontally. H for horizontal, V for vertical and B for both.